DS-889 Fix GraphExecutorClientSpec
There was a race condition in GraphExecutorClientSpec. GECA changes
its internal state when receives ExecutorSpawned from itself (after
requesting spawn on cluster). The test code has to send ExecutorReady
to the actor (but after ExecutorSpawned was already received). It did
not always happen causing GECA to drop the message.
From now, the test code sends ExecutorReady until it is handled.

Change-Id: I0d606b42c3e7ae23f3c033286654e9c1f012a9f5