User can connect to any cluster (Standalone, Yarn and Mesos)
WM_HOST and MQ_HOST no longer passed to SM by ENV. It's runtime property

We want frontend to hit proxy only. SM needs to be bound
to host network. Because of that proxy needs to be bound to
host network too. All other docker images have ports bound
so proxy has access to it.

Change-Id: I197f8e56a8eb50fc0358268552e445330986b215