Replace GatewayServer constructor hack with reflection
GatewayServer is missing a constructor we would like to use.
Previously, the problem was fixed with class duplication,
where our class has the constructor we need. It worked because
of correctly set classpath. On Yarn it does not work as expected
as the classpath is set externally and order of libraries
breaks the hack. The solution uses reflection and creates
the object manually.

Change-Id: I4f604e8f439f4987da4a3c21224d0b1067087d3b