DS-2060 Do not use mapValues collectReports()
MapValues method does not create a map (it's only a wrapper).
It means that, values will be calculated everytime you iterate
over a map created by mapValues(). This resulted in recalculation
of DOperables' reports. Moreover, a thread in WEA that should
not be blocked was blocked because of the recalculation and as
a result Session Executor was not responsive.

Change-Id: I2e9abd5605c13dd2add784f564b17c2e9700881e