DS-3208 Do not store node statuses in DB
Workflow Manager was storing not only results' ids and the corresponding
reports but also statuses of nodes. Storing statuses in the DB makes no sence,
as they to close to the execution process (imagine a situation where
the Executor goes down but we stored that a node is running).

The change includes also:
* deployment procedure changes (table's column in migration script),
* and minor changes in Session Manager integration tests to avoid
compile-time warnings.

Change-Id: I7e094ad8d7600b4cac7db6e8aae5176f0bfd0415