DS-2099 Multiple columns and in-place Transformer modes
This change introduces MultiColumnTransformer - a transformer that
can work on either a single column or multiple columns. Also, it can
also work in-place (by replacing columns) or not (new columns will
be appended to a DataFrame). When not working in-place and when
working with a single column one has to specify output column's name.
When working with multiple columns one has to speficy output column
names' prefix.

Other changes:
* UnaryTransformerAsMultiColumnTransformer - allows to wrap a Spark's
UnaryTransformer to create a MultiColumnTransformer.
* TypeConverter as a MultiColumnTransformer - TypeConverter rewritten
as a MultiColumnTransformer
* StringTokenizer as MultiColumnTransformer - example UnaryTransformer
from Spark as a MultiColumnTransformer.
* Small code-smells fixes

Change-Id: I0293daa8540e91a1b3f03b613cd931aad7e5fdda