Fix sessionmanager/build.sbt
The dot operator before the '!!' method meant that the command to be
executed is just the "cd..."  string, and not the whole Seq. Also, for
some magical reasons, build.sbt files need to have lazy, instead of
eager, vals - correcting this fixed the strange type errors that were
occuring whenever build.sbt was being recompiled in a non-clean
project dir.

Change-Id: I36b5eaa2fbf105decc64e9ab1c4f6bd027a24518