Update angular to 1.5 & DS-2905 Fix rendering problem resulting in merged components
Fix is applied attributes-serialized-view.js.
In angular 1.5 the the attributes-serialized-view was incorrectly rendered in wrong attribute-selector-type.
Initially it was rendered correctly, but while parsing ngIf directive, it was incorrectly moved to the wrong attribute-selector type.
I found few solutions to this problem:
- Remove replace:true from attributes-serialized-view so the whole element is wrapped with <attributes-serialized-view> element
- Wrap attributes-serialzied-view with div, so the top level element is not having ngIf directive
- Move attributes-serialized-view compilation to post link function of attributes-selector-type
- Replace ngIf directive from attributes-serialized-view with ngShow

Removal of replace:true was the most elegant solution.

Change-Id: I334e2f48f64cfff743eb1637c9e069ea5b0bb113