Change gc and decreased parallelism in local cluster
This is so that example 9 (credit cards) works better,
especially on machines with 8 cores.

Parallelism was set to maximum of 4. In local mode
we are limited by memory. When memory is limited,
too much partitions cause problems with GC.

GC policy was also changed to more effective one
- new policy lets example 9 complete on 8 cores,
but still it lasts really long because of GC.

Change-Id: I1d8623ed65e36532273458c89dce911ca7ef02b8