Fix authorization not creating admin bug
There was a bug that sometimes, admin was not created. It occured
when script checking if database is ready run at the moment when
schema_version table in uaa schema was created, but empty (not yet
populated). Reason for that was selecting something from
schema_version, which yields empty result if the query itself
is correct.

Change-Id: I18f90e4b3f056d19ba493299ceaa9f0472a0d8e6