SM-455 Kill spark using Spark Launcher handle. Improve error handling.
SM was initially created with Livy in mind. Livy was a service which would start
Spark Application. With Livy however we would not have easy way of communication
with Apache Spark process and not easy way of health checking application.

Back then we would communicate with App only over MQ. Health checking was to assume
that application broke if we would not get heartbeat over some period of time.
Killing was implemented by sending poison pill (over MQ) to the app.

Now, that we use Spark Launcher, we have better ways of killing and health
checking.

Change-Id: I44db51206fda31f0dcc732386382ed1a00b9f60a