Use double quotes to disambiguate sql column names
It turns out there are two different way of interpreting the
identifiers used in H2 sql statements depending on whether the
connection url has the DATABASE_TO_UPPER flag set to false or not. As
the column names in the schema creating script are explicitly wrapped
with double quotes, and hence are all lower case, to properly refer to
them without relying on this interpretation one needs to wrap them in
double quotes as well.

Change-Id: I8285459432dc74076eedcf3799a2bc67feb60ce6