Improve message error
Our Frontend dev would send requests with 'coma' instead of 'comma' in
nested 'csvParams' object which is required.

In such case object will not be parsed, and since it's optional, it
will be undefined in final parsed result.

It makes further validations fail, because of undefined object.

This change makes this explicit in error message. It follows convention
from swagger-generated code, where similar 'missing or invalid' message
is returned.

Change-Id: Iee906606fff81f9259a6239b8b6608c9b1e6dadd