Create repeatable migration script for workflow examples on each build.
To retain backwards compatibility with earlier versions of Seahorse I
restored the old V2__insert_examples.sql script, so that flyway won't
complain about checksums mismatch. However, the
generate_workflow_examples_sql.py script will produce a repeatable
(ie. not checksummed by flyway) migration that will overwrite the old
examples with whatever is currently in the repository. Thus, the
examples are no longer part of the flyway's database versioning
scheme, with the sole exception of the V2 version, which is now carved
in stone (unless we want to break the compatibility - in that case I'd
probably remove the V2 script altogether and renumber the rest).

Change-Id: Id91ed54bf25e37a72db1ca88fb0dfc23ccccb89c